Add express backend, JWT auth, profile icon logic
Added an `express` backend that uses `mongoose` to connect to a MongoDB Atlas instance.
The backend has two endpoints: `/register` and `/login`, both of which return a `JWT` token on a successful request.
Changed the profile icon to bring you to the `/profile` page when you are logged in and the `/login` page if you are not logged in.
implemented functional login and registration capabilities for the frontend with `axios.`