Refactor: secrets to .env, enhance user flow
Moved all of the secrets into a dedicated `.env` file.
Added an endpoint called `/user-details`, which allows to retrieve user details from the database based on a JWT.
Added links to the `/profile` page and the `/cart` page from the header.
Made data in the profile page be retrieved dynamically from the database.
Improved the CSS on the profile page.
Added a link to the registration page from the login page.