Add item model, split files, edit item details
Created a new `item` model for devices.
Split the `item` model and the `user` model into separate files.
Added an option to get `item` details and update them in the MongoDB database (Via `/item-details/:slug` and `/update-item/:slug`).
Fixed misspelling the `Register.js` file and some code quality issues there.
Added an actual option to edit `item` `stock` and `price`. When editing items, a sync button will appear, and when clicking it, the details will update in the database.