Use relative URLs, https, add pm2 config
Updated the files to use the relative `URL` to the current `hostname` instead of localhost to allow hosting on different webservers.
Changed the protocol on one of the fetch requests to use `https` instead of `http`.
Added an `ecosystem.config.js` file for the capability of running a production server with `pm2`.